requirements.txt
antiorm==1.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
db==0.1.1
db-sqlite3==0.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
greenlet==3.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
peewee==3.17.0
pluggy==1.3.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
tomli==2.0.1
typer==0.9.0
typing_extensions==4.8.0
